Fixed bug related to functionality of HashMap and identical characters
For strings like "banana" and "ban", the "ana" will overwrite the original 'a' and 'n', resulting in an identical HashMap to "ban" even though ban is not an anagram of banana. To fix the issue I added a conditional that tests if the lengths of the two strings are identical because if they are not, then an anagram cannot be created.